rename utils.py to avoid conflict with cldbe utils/utils.py
