+++ tests/test_extract_entities.py
@@ -1,7 +1,7 @@
 import pytest
 import neointerface
 import pandas as pd
-from utils.utils import compare_recordsets
+from utils.neointerface_utils import compare_recordsets
 
 
 # Provide a database connection that can be used by the various tests that need it
